+++ modules/core/dependency/python-ihm/ChangeLog.rst
@@ -1,3 +1,11 @@
+0.25 - 2021-12-03
+=================
+  - :func:`ihm.dictionary.Dictionary.validate` will now report errors for
+    any keywords or categories in the file that are not present in the
+    dictionary.
+  - :class:`ihm.LPeptideAlphabet` now supports the ASX and GLX ambiguous
+    residue types.
+
 0.24 - 2021-12-01
 =================
   - :class:`ihm.AsymUnit` now supports insertion codes in its